requirement.txt
Flask==2.3.2
Flask-Session==0.5.0
openai==0.27.0
python-dotenv==1.0.0
pydub==0.25.1
SoundFile==0.10.3.post1
Werkzeug==2.3.3
